requirements.txt
djangorestframework
markdown
django-filter
drf-yasg
faker>=18.9.0
pytest